backend/requirements.txt
langchain_core
langchain_openai
langchain_community
langchain-qdrant
qdrant-client
langsmith
openai
tiktoken
cohere
lxml